pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic-retrieval-engine"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "faiss-cpu>=1.8.0",
    "python-docx>=1.1",
    "pypdf>=4.0",
    "python-magic>=0.4",
    "sentence-transformers>=2.7",
    "pillow>=10.0"
]

[project.optional-dependencies]
dev = [
    "pytest"
]

[tool.setuptools.packages.find]
where = ["src"]

#suppress weird deprecation warnings from using embedding model
[tool.pytest.ini_options]
pythonpath = ["src"]
filterwarnings = [
    "ignore:.*SwigPy.*:DeprecationWarning",
    "ignore:.*swigvarlink.*:DeprecationWarning"
]